Optimization and stuff hamdouallah

+++ src/main/java/Controllers/FilmController.java
@@ -39,7 +39,7 @@ public class FilmController {
     public static List<Film> getMostRecentFilm(int numSeries) throws SQLException, IOException{
         return FilmService.getMostRecentFilm(numSeries);
     }
-    public static Long getscorepourcantage(Film film) {
+    public static double getscorepourcantage(Film film) {
         return FilmService.getscorepourcantage(film);
     }
